FINALLY SOLVED THE TREE CLONE ISSUE.
Since every device has to implement clone, if it's a connectable device
it will need a check for clone arguments, otherwise the argument is
unused. This allows all devices to be cloned, but also trees to be
cloned with "nodes" being defined by objects that do not have a check
for clone arguments. bleh.